Fixed quite a few things.
In LeadsToWar.html, I just cleaned up some code that was bothering me
(because of spacing), and added an Attack Opposing Team button. Still
trying to get this to work fully, but right now, it kind of works.

In LeadsToWar.js, I added the team property. This allows us to, just by
looking at the different objects in the array, see what team they are
on.

In LeadsToWarFunctions.js, I modified the attackUnit function a little
bit, just so that it fit in with the button.